Disable likes for article authors

+++ src/utils/helpers.js
@@ -2,6 +2,6 @@ export const formatDateTimeString = (dateTimeStr) => {
   const date = new Date(dateTimeStr);
   return Intl.DateTimeFormat('en-GB', {
     dateStyle: 'full',
-    timeStyle: 'long',
+    timeStyle: 'short',
   }).format(date);
 };
